refactor: move react specific code to features
