+++ src/features/NoteEditor/index.tsx
@@ -31,8 +31,10 @@ import { SuggestedTagsList } from '@components/SuggestedTagsList';
 import { findLinksInText, getResourceIdInUrl } from '@core/features/links';
 import { INote, INoteContent } from '@core/features/notes';
 import { IResolvedTag } from '@core/features/tags';
-import { useTagsContext } from '@state/tags';
-import { useWorkspaceContext } from '@state/workspace';
+import {
+	useTagsContext,
+	useWorkspaceContext,
+} from '@features/Workspace/WorkspaceProvider';
 
 import { FileUploader } from '../MonakoEditor/features/useDropFiles';
 import { MonacoEditor } from '../MonakoEditor/MonacoEditor';
